m_12_FastAPI_OAuth_project/requirements.txt
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
Authlib==1.3.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
environs==11.0.0
fastapi==0.110.0
fastapi-jwt-auth==0.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
libgravatar==1.0.4
MarkupSafe==2.1.5
marshmallow==3.21.1
packaging==24.0
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==1.7.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.29.0